Detects number of channels in a sample and plays it Mono, Stereo Left, Stereo Right, Stereo Sum, depending on the stereo-mode switch setting. Only works with 16-bit wav files at this point. Added trigger_jacks files. Updated Makefile to use arm-none-eabi-gcc version 5.4. Worked around the bug that makes the chip go into an infinite loop when dividing 64-bit integers. This might be some sort of linking issue on my part...? The "bug" was "fixed" by using 32-bit division if possible and float division if not. Updated system_stm32f4xx.c to run at 44.1kHz Still have major issues: cannot play two files at a high speed, there's a noticable lag between trigger and audio starting, sometimes doesn't read the SDCARD on boot, can't re-trigger quickly or else it gets an error...

+++ inc/fatfs/integer.h
@@ -33,7 +33,8 @@ typedef int32_t			LONG;
 typedef uint32_t		DWORD;
 
 /* This type MUST be 64-bit (Remove this for C89 compatibility) */
-typedef unsigned long long QWORD;
+//typedef unsigned long long QWORD;
+typedef uint64_t QWORD;
 
 #endif
